requirements.txt
Flask>=2.3
tensorflow
python-dotenv
requests
geopy 
scikit-learn
pandas
colorama
aiohttp